Rspec done
