+++ app/controllers/users/registrations_controller.rb
@@ -2,8 +2,8 @@
 
 module Users
   class RegistrationsController < Devise::RegistrationsController
-    # before_action :configure_sign_up_params, only: [:create]
-    # before_action :configure_account_update_params, only: [:update]
+    before_action :configure_sign_up_params, only: [:create]
+    before_action :configure_account_update_params, only: [:update]
 
     # GET /resource/sign_up
     # def new
@@ -11,9 +11,9 @@ module Users
     # end
 
     # POST /resource
-    # def create
-    #   super
-    # end
+    def create
+      super
+    end
 
     # GET /resource/edit
     # def edit
@@ -21,9 +21,9 @@ module Users
     # end
 
     # PUT /resource
-    # def update
-    #   super
-    # end
+    def update
+      super
+    end
 
     # DELETE /resource
     # def destroy
@@ -42,14 +42,18 @@ module Users
     # protected
 
     # If you have extra params to permit, append them to the sanitizer.
-    # def configure_sign_up_params
-    #   devise_parameter_sanitizer.permit(:sign_up, keys: [])
-    # end
+    def configure_sign_up_params
+      devise_parameter_sanitizer.permit(
+        :sign_up, keys: %i[fullname]
+      )
+    end
 
     # If you have extra params to permit, append them to the sanitizer.
-    # def configure_account_update_params
-    #   devise_parameter_sanitizer.permit(:account_update, keys: [])
-    # end
+    def configure_account_update_params
+      devise_parameter_sanitizer.permit(
+        :account_update, keys: %i[fullname]
+      )
+    end
 
     # The path used after sign up.
     # def after_sign_up_path_for(resource)
