+++ app/controllers/users/sessions_controller.rb
@@ -2,21 +2,17 @@
 
 module Users
   class SessionsController < Devise::SessionsController
-    # before_action :configure_sign_in_params, only: [:create]
+    before_action :configure_sign_up_params, only: [:create]
 
     # GET /resource/sign_in
-    def new
-      super
-    end
+    # def new
+    #   super
+    # end
 
     # POST /resource/sign_in
-    # def create
-    #  self.resource = warden.authenticate!(auth_options)
-    #  set_flash_message!(:notice, :signed_in)
-    #  sign_in(resource_name, resource)
-    #  yield resource if block_given?
-    #  respond_with resource, location: after_sign_in_path_for(resource)
-    # end
+    def create
+      super
+    end
 
     # DELETE /resource/sign_out
     # def destroy
@@ -26,8 +22,10 @@ module Users
     # protected
 
     # If you have extra params to permit, append them to the sanitizer.
-    # def configure_sign_in_params
-    #   devise_parameter_sanitizer.permit(:sign_in, keys: [:attribute])
-    # end
+    def configure_sign_up_params
+      devise_parameter_sanitizer.permit(
+        :sign_up, keys: %i[fullname]
+      )
+    end
   end
 end
